refactor(@angular-devkit/build-angular): move dev-server specs into a seperate directory
(cherry picked from commit ef753628a648f4b15ce11531e879b9bab73b7385)
