+++ packages/angular_devkit/build_angular/src/builders/dev-server/specs/hmr_spec.ts
@@ -13,7 +13,7 @@ import { Page } from 'puppeteer/lib/cjs/puppeteer/common/Page';
 import puppeteer from 'puppeteer/lib/cjs/puppeteer/node';
 /* eslint-enable import/no-extraneous-dependencies */
 import { debounceTime, switchMap, take } from 'rxjs/operators';
-import { createArchitect, host } from '../../testing/test-utils';
+import { createArchitect, host } from '../../../testing/test-utils';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 declare const document: any;
