+++ packages/angular_devkit/build_angular/src/builders/dev-server/specs/live-reload_spec.ts
@@ -14,7 +14,7 @@ import { Browser } from 'puppeteer/lib/cjs/puppeteer/common/Browser';
 import { Page } from 'puppeteer/lib/cjs/puppeteer/common/Page';
 import puppeteer from 'puppeteer/lib/cjs/puppeteer/node';
 import { debounceTime, switchMap, take } from 'rxjs/operators';
-import { createArchitect, host } from '../../testing/test-utils';
+import { createArchitect, host } from '../../../testing/test-utils';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 declare const document: any;
