+++ packages/angular_devkit/build_angular/src/builders/dev-server/specs/works_spec.ts
@@ -10,7 +10,7 @@ import { Architect, BuilderRun } from '@angular-devkit/architect';
 import { DevServerBuilderOutput } from '@angular-devkit/build-angular';
 import { normalize, virtualFs } from '@angular-devkit/core';
 import fetch from 'node-fetch'; // eslint-disable-line import/no-extraneous-dependencies
-import { createArchitect, host } from '../../testing/test-utils';
+import { createArchitect, host } from '../../../testing/test-utils';
 
 describe('Dev Server Builder', () => {
   const target = { project: 'app', target: 'serve' };
